+++ documentation/changelog.md
@@ -1,13 +1,16 @@
 ## TODO:
 - Fix sound still playing after canvasser closed/reloaded/loadinto.
 - Load/Save File manage to Box/Dropbox/Google Drive.
-- Shape copy
 - Load Autoback
 - Fade shapes
 - Tooltip/Instructions to use variable in text shape with {{ var_name }}
-
+- Can't remove colors from shape object
+- Can't delete shape drawcode
 
 ## Changelog:
+- 2018-08-02 Added setting for responsive Canvasser.
+- 2018-07-31 Added copy to Shape menu.
+- 2018-07-31 Fixed conditionals not working.
 - 2018-07-26 Added text entry through DOM to allow variables values to be changed.
 - 2018-07-17 Make sure an anim is not undefined before trying to pause it.
 - 2018-07-11 Markdown like format for help docs.